Update the prepare script to build prepared data in memory
While it is generally a fine practice to write out to a file on the local
machine for intermediary data, it can be advantageous to do everything in
memory (in particular, the 1st generation Google Cloud Functions platform runs
scripts within a read-only file system).

There are several ways to avoid writing to the file system in Python (e.g.,
using a io.StringIO file-like object), but in this commit we elect to just add
each row to a list, and then join all the rows in that list with newline
characters. Then we use upload_from_string instead of upload_from_file to send
the data to Cloud Storage.